sync to latest

+++ server/repo.go
@@ -1,14 +1,28 @@
+// Copyright 2018 Drone.IO Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package server
 
 import (
 	"encoding/base32"
 	"fmt"
 	"net/http"
+	"strconv"
 
 	"github.com/gin-gonic/gin"
 	"github.com/gorilla/securecookie"
 
-	"github.com/drone/drone/cache"
 	"github.com/drone/drone/model"
 	"github.com/drone/drone/remote"
 	"github.com/drone/drone/router/middleware/session"
@@ -20,50 +34,45 @@ import (
 func PostRepo(c *gin.Context) {
 	remote := remote.FromContext(c)
 	user := session.User(c)
-	owner := c.Param("owner")
-	name := c.Param("name")
+	repo := session.Repo(c)
 
-	if user == nil {
-		c.AbortWithStatus(403)
+	if repo.IsActive {
+		c.String(409, "Repository is already active.")
 		return
 	}
 
-	r, err := remote.Repo(user, owner, name)
-	if err != nil {
-		c.String(404, err.Error())
-		return
-	}
-	m, err := cache.GetPerms(c, user, owner, name)
-	if err != nil {
-		c.String(404, err.Error())
-		return
-	}
-	if !m.Admin {
-		c.String(403, "Administrative access is required.")
+	if err := Config.Services.Limiter.LimitRepo(user, repo); err != nil {
+		c.String(403, "Repository activation blocked by limiter")
 		return
 	}
 
-	// error if the repository already exists
-	_, err = store.GetRepoOwnerName(c, owner, name)
-	if err == nil {
-		c.String(409, "Repository already exists.")
-		return
+	repo.IsActive = true
+	repo.UserID = user.ID
+	if !repo.AllowPush && !repo.AllowPull && !repo.AllowDeploy && !repo.AllowTag {
+		repo.AllowPush = true
+		repo.AllowPull = true
+	}
+	if repo.Visibility == "" {
+		repo.Visibility = model.VisibilityPublic
+		if repo.IsPrivate {
+			repo.Visibility = model.VisibilityPrivate
+		}
+	}
+	if repo.Config == "" {
+		repo.Config = Config.Server.RepoConfig
+	}
+	if repo.Timeout == 0 {
+		repo.Timeout = 60 // 1 hour default build time
+	}
+	if repo.Hash == "" {
+		repo.Hash = base32.StdEncoding.EncodeToString(
+			securecookie.GenerateRandomKey(32),
+		)
 	}
 
-	// set the repository owner to the
-	// currently authenticated user.
-	r.UserID = user.ID
-	r.AllowPush = true
-	r.AllowPull = true
-	r.Config = ".drone.yml"
-	r.Timeout = 60 // 1 hour default build time
-	r.Hash = base32.StdEncoding.EncodeToString(
-		securecookie.GenerateRandomKey(32),
-	)
-
-	// crates the jwt token used to verify the repository
-	t := token.New(token.HookToken, r.FullName)
-	sig, err := t.Sign(r.Hash)
+	// creates the jwt token used to verify the repository
+	t := token.New(token.HookToken, repo.FullName)
+	sig, err := t.Sign(repo.Hash)
 	if err != nil {
 		c.String(500, err.Error())
 		return
@@ -75,22 +84,24 @@ func PostRepo(c *gin.Context) {
 		sig,
 	)
 
-	// activate the repository before we make any
-	// local changes to the database.
-	err = remote.Activate(user, r, link)
+	err = remote.Activate(user, repo, link)
 	if err != nil {
 		c.String(500, err.Error())
 		return
 	}
 
-	// persist the repository
-	err = store.CreateRepo(c, r)
+	from, err := remote.Repo(user, repo.Owner, repo.Name)
+	if err == nil {
+		repo.Update(from)
+	}
+
+	err = store.UpdateRepo(c, repo)
 	if err != nil {
 		c.String(500, err.Error())
 		return
 	}
 
-	c.JSON(200, r)
+	c.JSON(200, repo)
 }
 
 func PatchRepo(c *gin.Context) {
@@ -132,6 +143,18 @@ func PatchRepo(c *gin.Context) {
 	if in.Config != nil {
 		repo.Config = *in.Config
 	}
+	if in.Visibility != nil {
+		switch *in.Visibility {
+		case model.VisibilityInternal, model.VisibilityPrivate, model.VisibilityPublic:
+			repo.Visibility = *in.Visibility
+		default:
+			c.String(400, "Invalid visibility type")
+			return
+		}
+	}
+	if in.BuildCounter != nil {
+		repo.Counter = *in.BuildCounter
+	}
 
 	err := store.UpdateRepo(c, repo)
 	if err != nil {
@@ -160,18 +183,30 @@ func GetRepo(c *gin.Context) {
 }
 
 func DeleteRepo(c *gin.Context) {
+	remove, _ := strconv.ParseBool(c.Query("remove"))
 	remote := remote.FromContext(c)
 	repo := session.Repo(c)
 	user := session.User(c)
 
-	err := store.DeleteRepo(c, repo)
+	repo.IsActive = false
+	repo.UserID = 0
+
+	err := store.UpdateRepo(c, repo)
 	if err != nil {
 		c.AbortWithError(http.StatusInternalServerError, err)
 		return
 	}
 
+	if remove {
+		err := store.DeleteRepo(c, repo)
+		if err != nil {
+			c.AbortWithError(http.StatusInternalServerError, err)
+			return
+		}
+	}
+
 	remote.Deactivate(user, repo, httputil.GetURL(c.Request))
-	c.Writer.WriteHeader(http.StatusOK)
+	c.JSON(200, repo)
 }
 
 func RepairRepo(c *gin.Context) {
@@ -179,7 +214,93 @@ func RepairRepo(c *gin.Context) {
 	repo := session.Repo(c)
 	user := session.User(c)
 
-	// crates the jwt token used to verify the repository
+	// creates the jwt token used to verify the repository
+	t := token.New(token.HookToken, repo.FullName)
+	sig, err := t.Sign(repo.Hash)
+	if err != nil {
+		c.String(500, err.Error())
+		return
+	}
+
+	// reconstruct the link
+	host := httputil.GetURL(c.Request)
+	link := fmt.Sprintf(
+		"%s/hook?access_token=%s",
+		host,
+		sig,
+	)
+
+	remote.Deactivate(user, repo, host)
+	err = remote.Activate(user, repo, link)
+	if err != nil {
+		c.String(500, err.Error())
+		return
+	}
+
+	from, err := remote.Repo(user, repo.Owner, repo.Name)
+	if err == nil {
+		repo.Name = from.Name
+		repo.Owner = from.Owner
+		repo.FullName = from.FullName
+		repo.Avatar = from.Avatar
+		repo.Link = from.Link
+		repo.Clone = from.Clone
+		repo.IsPrivate = from.IsPrivate
+		if repo.IsPrivate != from.IsPrivate {
+			repo.ResetVisibility()
+		}
+		store.UpdateRepo(c, repo)
+	}
+
+	c.Writer.WriteHeader(http.StatusOK)
+}
+
+func MoveRepo(c *gin.Context) {
+	remote := remote.FromContext(c)
+	repo := session.Repo(c)
+	user := session.User(c)
+
+	to, exists := c.GetQuery("to")
+	if !exists {
+		err := fmt.Errorf("Missing required to query value")
+		c.AbortWithError(http.StatusInternalServerError, err)
+		return
+	}
+
+	owner, name, errParse := model.ParseRepo(to)
+	if errParse != nil {
+		c.AbortWithError(http.StatusInternalServerError, errParse)
+		return
+	}
+
+	from, err := remote.Repo(user, owner, name)
+	if err != nil {
+		c.AbortWithError(http.StatusInternalServerError, err)
+		return
+	}
+	if !from.Perm.Admin {
+		c.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+
+	repo.Name = from.Name
+	repo.Owner = from.Owner
+	repo.FullName = from.FullName
+	repo.Avatar = from.Avatar
+	repo.Link = from.Link
+	repo.Clone = from.Clone
+	repo.IsPrivate = from.IsPrivate
+	if repo.IsPrivate != from.IsPrivate {
+		repo.ResetVisibility()
+	}
+
+	errStore := store.UpdateRepo(c, repo)
+	if errStore != nil {
+		c.AbortWithError(http.StatusInternalServerError, errStore)
+		return
+	}
+
+	// creates the jwt token used to verify the repository
 	t := token.New(token.HookToken, repo.FullName)
 	sig, err := t.Sign(repo.Hash)
 	if err != nil {
